pyproject.toml
[tool.poetry]
name = "learning-fastapi"
version = "0.1.0"
description = ""
authors = ["Kumar Pratik"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
cookiecutter = "^2.1.1"
gunicorn = "^20.1.0"
Jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.37"
pytest = "^7.1.2"
pydantic = "^1.9.1"
alembic = "^1.8.0"
python-dotenv = "^0.20.0"
sqlmodel = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
